[Fix] lhasa: Apply commit <https://github.com/fragglet/lhasa/commit/59c43fc841bfdba970a810e21d97f5e074cd286e>.
git-svn-id: https://source.openmpt.org/svn/openmpt/trunk/OpenMPT@9336 56274372-70c3-4bfc-bfc3-4c3a0b034d27

+++ include/lhasa/lib/lha_endian.c
@@ -27,8 +27,10 @@ uint16_t lha_decode_uint16(uint8_t *buf)
 
 uint32_t lha_decode_uint32(uint8_t *buf)
 {
-	return (uint32_t) (buf[0] | (buf[1] << 8)
-	                 | (buf[2] << 16) | (buf[3] << 24));
+	return ((uint32_t) buf[0])
+	     | ((uint32_t) buf[1] << 8)
+	     | ((uint32_t) buf[2] << 16)
+	     | ((uint32_t) buf[3] << 24);
 }
 
 uint64_t lha_decode_uint64(uint8_t *buf)
@@ -50,7 +52,9 @@ uint16_t lha_decode_be_uint16(uint8_t *buf)
 
 uint32_t lha_decode_be_uint32(uint8_t *buf)
 {
-	return (uint32_t) ((buf[0] << 24) | (buf[1] << 16)
-	                 | (buf[2] << 8) | buf[3]);
+	return ((uint32_t) buf[0] << 24)
+	     | ((uint32_t) buf[1] << 16)
+	     | ((uint32_t) buf[2] << 8)
+	     | ((uint32_t) buf[3]);
 }
